Add Pico OXR build variant
This commit introduces a new build variant "picooxr" using Pico's OpenXR
SDK, and makes necessary changes to get Wolvic to run on the Pico 4. It
is likely that the variant also works with at least some versions of the
Pico Neo 3, but that was not tested.

In terms of limitations, WebXR does not seem to work correctly and this
commit does not contain images or 3D models of the Pico 4 controllers.
For controllers, official models are available as part of the Pico SDKs,
but it is unclear if they are license-compatible with this repository.